chore: clean F2018 intrinsic validator and fixtures

+++ tests/fixtures/Fortran2018/test_issue194_intrinsic_semantics/collective_function_intrinsics.f90
@@ -1,8 +1,8 @@
 program collective_function_intrinsics
-    use, intrinsic :: iso_fortran_env, only: team_type
+    use, intrinsic :: iso_fortran_env, only: team_type, dp => real64
     implicit none
     integer :: codims(2), tnum
-    real :: x[*]
+    real(dp) :: x[*]
     type(team_type) :: t
 
     codims = coshape(x)
