+++ tests/fixtures/Fortran2018/test_issue194_intrinsic_semantics/random_init_intrinsic.f90
@@ -1,6 +1,7 @@
 program random_init_intrinsic
+    use, intrinsic :: iso_fortran_env, only: dp => real64
     implicit none
-    real :: x
+    real(dp) :: x
 
     call random_init(repeatable=.true., image_distinct=.false.)
     call random_number(x)
